Rename nano/common.hpp to nano/endpoint.hpp and extract endpoint_templ.hpp template implementation file.
