+++ nano/core_test/fakes/work_peer.hpp
@@ -4,7 +4,7 @@
 #include <nano/lib/locks.hpp>
 #include <nano/lib/numbers.hpp>
 #include <nano/lib/work.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 
 #include <boost/asio.hpp>
 #include <boost/asio/steady_timer.hpp>
