+++ nano/core_test/network_filter.cpp
@@ -1,7 +1,7 @@
 #include <nano/lib/blocks.hpp>
 #include <nano/lib/network_filter.hpp>
 #include <nano/lib/stream.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 #include <nano/secure/common.hpp>
 #include <nano/test_common/testutil.hpp>
