+++ nano/core_test/voting.cpp
@@ -1,5 +1,5 @@
 #include <nano/lib/blocks.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 #include <nano/node/local_vote_history.hpp>
 #include <nano/node/vote_generator.hpp>
 #include <nano/node/vote_spacing.hpp>
