+++ nano/node/bootstrap/bootstrap_bulk_push.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 
 #include <future>
