+++ nano/node/bootstrap/bootstrap_connections.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <nano/node/bootstrap/bootstrap_bulk_pull.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 #include <nano/node/transport/tcp_socket.hpp>
 
 #include <atomic>
