+++ nano/node/network.hpp
@@ -2,7 +2,7 @@
 
 #include <nano/lib/logging.hpp>
 #include <nano/lib/network_filter.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 #include <nano/node/messages.hpp>
 #include <nano/node/peer_exclusion.hpp>
 #include <nano/node/transport/common.hpp>
