+++ nano/node/transport/transport.hpp
@@ -3,7 +3,7 @@
 #include <nano/lib/locks.hpp>
 #include <nano/lib/stats.hpp>
 #include <nano/node/bandwidth_limiter.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 #include <nano/node/messages.hpp>
 #include <nano/node/transport/tcp_socket.hpp>
