+++ nano/rpc_test/entry.cpp
@@ -1,6 +1,6 @@
 #include <nano/lib/logging.hpp>
 #include <nano/lib/memory.hpp>
-#include <nano/node/common.hpp>
+#include <nano/node/endpoint.hpp>
 
 #include <gtest/gtest.h>
